+++ apps/ymir-sdl3/src/main.cpp
@@ -10,7 +10,7 @@ int main(int argc, char **argv) {
 
     app::CommandLineOptions progOpts{};
     cxxopts::Options options("Ymir", "Ymir - Sega Saturn emulator");
-    options.add_options()("d,disc", "Path to Saturn disc image (.ccd, .cue, .iso, .mds)",
+    options.add_options()("d,disc", "Path to Saturn disc image (.ccd, .chd, .cue, .iso, .mds)",
                           cxxopts::value(progOpts.gameDiscPath));
     options.add_options()("p,profile", "Path to profile directory", cxxopts::value(progOpts.profilePath));
     options.add_options()("h,help", "Display help text", cxxopts::value(showHelp)->default_value("false"));