+++ libs/ymir-core/include/ymir/media/loader/loader.hpp
@@ -1,18 +1,13 @@
 #pragma once
 
-#include "loader_bin_cue.hpp"
-#include "loader_img_ccd_sub.hpp"
-#include "loader_iso.hpp"
-#include "loader_mdf_mds.hpp"
-
 #include <ymir/media/disc.hpp>
 
 #include <filesystem>
-#include <vector>
 
 namespace ymir::media {
 
 // Attempts to load a CD image from any of the supported file formats:
+//   MAME CHD    (if provided a .chd file)
 //   BIN/CUE     (if provided a .cue file)
 //   MDF/MDS     (if provided a .mds file)
 //   IMG/CCD/SUB (if provided a .ccd file)