+++ libs/ymir-core/src/ymir/media/loader/loader.cpp
@@ -1,22 +1,21 @@
 #include <ymir/media/loader/loader.hpp>
 
+#include <ymir/media/loader/loader_bin_cue.hpp>
+#include <ymir/media/loader/loader_chd.hpp>
+#include <ymir/media/loader/loader_img_ccd_sub.hpp>
+#include <ymir/media/loader/loader_iso.hpp>
+#include <ymir/media/loader/loader_mdf_mds.hpp>
+
 namespace ymir::media {
 
 bool LoadDisc(std::filesystem::path path, Disc &disc, bool preloadToRAM) {
-    if (loader::bincue::Load(path, disc, preloadToRAM)) {
-        return true;
-    }
-    if (loader::mdfmds::Load(path, disc, preloadToRAM)) {
-        return true;
-    }
-    if (loader::ccd::Load(path, disc, preloadToRAM)) {
-        return true;
-    }
-    // NOTE: ISO must be the last to be tested since its detection is more lenient
-    if (loader::iso::Load(path, disc, preloadToRAM)) {
-        return true;
-    }
-    return false;
+    // Abuse short-circuiting to pick the first matching loader with less verbosity
+    return loader::chd::Load(path, disc, preloadToRAM) ||    //
+           loader::bincue::Load(path, disc, preloadToRAM) || //
+           loader::mdfmds::Load(path, disc, preloadToRAM) || //
+           loader::ccd::Load(path, disc, preloadToRAM) ||    //
+           // NOTE: ISO must be the last to be tested since its detection is more lenient
+           loader::iso::Load(path, disc, preloadToRAM);
 }
 
 } // namespace ymir::media